% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_missing.R
\name{filter_missing}
\alias{filter_missing}
\title{Filter the missing values with IDs}
\usage{
filter_missing(conn, data)
}
\arguments{
\item{conn}{A expression like x ~ id, x is the numerical variable.}

\item{data}{A data frame, data frame extension (e.g. a tibble).}
}
\value{
A data frame
}
\description{
Filter the missing values with IDs
}
\examples{
conn <- x ~ id
data <- data.frame(
  id = c(1:6),
  x = c(5.2, NA, 3.9, NA, 4.5, 4.3)
)
filter_missing(conn, data)
}
