% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_outlier.R
\name{filter_outlier}
\alias{filter_outlier}
\title{Filter the outlies with IDs}
\usage{
filter_outlier(conn, data, coef = 1.5)
}
\arguments{
\item{conn}{A expression like x ~ id, x is the numerical variable.}

\item{data}{A data frame, data frame extension (e.g. a tibble).}

\item{coef}{A numerical value. Defaults to 1.5.}
}
\value{
A data frame
}
\description{
Filter the outlies with IDs
}
\examples{
conn <- x ~ id
data <- data.frame(
  id = c(1:6),
  x = c(5.2, 4.5, 3.9, 30, 4.5, 4.3))
filter_outlier(conn, data)
filter_outlier(conn, data, coef = 1.5)
}
