% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_3line_table.R
\name{flex_3line_table}
\alias{flex_3line_table}
\title{Create a 3 line flextable}
\usage{
flex_3line_table(x)
}
\arguments{
\item{x}{A flextable object.}
}
\value{
A flextable object.
}
\description{
Create a 3 line flextable
}
\examples{
library(flextable)
x <- data.frame(
  sex = c("F", "M", "F"),
  age = c(22, 20, 30))
x <- flextable(x)
flex_3line_table(x)
}
