% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_tibble.R
\name{list_to_tibble}
\alias{list_to_tibble}
\title{Convert list-column data frame (including different lengths) to a common one}
\usage{
list_to_tibble(x)
}
\arguments{
\item{x}{A list-columns}
}
\value{
A tibble data frame
}
\description{
Convert list-column data frame (including different lengths) to a common one
}
\examples{
library(tibble)
x <- tibble(sex = list(c("F", "M", "M")), age = list(c(22, 20, 23)))
list_to_tibble(x)
}
